controller/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
kubernetes
requests>=2.31.0
pyyaml>=6.0